Run as 64 bit app (so can also hook modern hitman games in principle, for future proofing).

+++ base_ptrs.hpp
@@ -3,4 +3,4 @@
 #include <array>
 #include <cstdint>
 
-using BasePtrs = std::array<int32_t, 5>;  // up to 5 module base pointers
+using BasePtrs = std::array<intptr_t, 5>;  // up to 5 module base pointers
