+++ mem/alloc.hpp
@@ -6,7 +6,7 @@
 
 struct Alloc {
     std::shared_ptr<void> handle;
-    int32_t ptr;
+    intptr_t ptr;
 };
 
 struct AllocDeleter {
@@ -15,4 +15,4 @@ struct AllocDeleter {
 
 using AllocPtr = std::unique_ptr<Alloc, AllocDeleter>;
 
-AllocPtr virtual_alloc_ex(std::shared_ptr<void> handle, int32_t size);
+AllocPtr virtual_alloc_ex(std::shared_ptr<void> handle, ptrdiff_t size);
